Make Unit/Measure as a domain.
