+++ src/features/suggestions/parser/amount.test.ts
@@ -1,5 +1,6 @@
 import _ from "lodash";
-import { parseAmount, Unit } from "./amount";
+import { parseAmount } from "./amount";
+import { Unit } from "../../../model/Unit";
 
 test("unit pound(s) -> lb", () => {
   testCases("pound", "lb").run();
