+++ src/features/suggestions/search/isConvertible.ts
@@ -1,6 +1,7 @@
 import _ from 'lodash';
 import { PredefinedSuggestion } from './search';
-import { isMeasure, measureOf, parseAmount, Unit } from '../parser/amount';
+import { parseAmount } from '../parser/amount';
+import { isMeasure, measureOf, Unit } from "../../../model/Unit";
 
 export default function isConvertible(fromUnit: Unit, suggestion: PredefinedSuggestion) {
   if (_.isUndefined(fromUnit))
